DOSIS-12941 Refactoring code and tests
