+++ tests/service_automation/tests/utilities/infra/s3_util.py
@@ -5,7 +5,6 @@ from loguru import logger
 
 class S3Utils:
     def __init__(self):
-        """Initialize AWS S3 client using credentials from the terminal (AWS CLI)"""
         try:
             # Use default AWS CLI profile and automatically detect region
             session = boto3.Session()
@@ -16,7 +15,6 @@ class S3Utils:
             raise
 
     def list_buckets(self):
-        """Returns a list of S3 bucket names."""
         try:
             response = self.s3_client.list_buckets()
             return [bucket["Name"] for bucket in response["Buckets"]]
@@ -25,10 +23,6 @@ class S3Utils:
             return []
 
     def check_bucket_exists(self, bucket_name):
-        """
-        Determine whether the bucket exists and you have access to it.
-        :return: True when the bucket exists; otherwise, False.
-        """
         try:
             self.s3_client.head_bucket(Bucket=bucket_name)
             exists = True
@@ -38,11 +32,6 @@ class S3Utils:
         return exists
 
     def get_bucket(self, project, workspace, env, stack, bucket):
-        """
-        Structures the full s3 bucket name based on the project, workspace, environment, stack
-        and bucket.
-        :return: The full bucket name.
-        """
         logger.debug(f"project: {project}, bucket: {bucket}, stack: {stack}, env: {env}, workspace: {workspace}")
         if workspace == "":
             bucket_name = project + "-" + env + "-" + stack + "-" + bucket
@@ -51,10 +40,6 @@ class S3Utils:
         return bucket_name
 
     def get_object(self, bucket_name, filename):
-        """
-        Connects to the S3 bucket, retrieves the file and returns the file data
-        :return: The csv file data.
-        """
         try:
             response = self.s3_client.get_object(
                 Bucket=bucket_name,
@@ -66,27 +51,18 @@ class S3Utils:
         return file_data
 
     def download_object(self, bucket_name, filepath, filename):
-        """
-        Connects to the S3 bucket and downloads the file
-        """
         try:
             self.s3_client.download_file(Bucket=bucket_name, Key=filename, Filename=filepath)
         except Exception:
             logger.error("Error: file could not be downloaded")
 
     def put_object(self, bucket_name, filepath, file_name):
-        """
-        Connects to the S3 bucket and uploads the file
-        """
         try:
             self.s3_client.upload_file(Filename=filepath, Bucket=bucket_name, Key=file_name)
         except Exception:
             logger.error("Error: file could not be uploaded")
 
     def delete_object(self, bucket_name, filename):
-        """"
-        Connects to the S3 bucket and uploads the file
-        """
         try:
             self.s3_client.delete_object(Bucket=bucket_name, Key=filename)
         except Exception:
